pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "~3.11"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
